refactor: extract state module tests
